requirements.txt
numpy
tqdm
sentence-transformers
scikit-learn
python-dateutil
pypdf
beautifulsoup4

opencv-python
Pillow
imagehash

openai-whisper
ffmpeg-python
easyocr

torch
transformers
